+++ tick/optim/model/src/hawkes_fixed_expkern_loglik.cpp
@@ -1,5 +1,6 @@
 
 #include "hawkes_fixed_expkern_loglik.h"
+#include "parallel/parallel.h"
 
 ModelHawkesFixedExpKernLogLik::ModelHawkesFixedExpKernLogLik(
     const double decay, const int max_n_threads) :
@@ -92,7 +93,7 @@ void ModelHawkesFixedExpKernLogLik::grad(const ArrayDouble &coeffs,
 
   // This allows to run in a multithreaded environment the computation of each component
   parallel_run(get_n_threads(), n_nodes, &ModelHawkesFixedExpKernLogLik::grad_dim_i, this, coeffs, out);
-  out /= n_total_jumps;
+  out /= static_cast<double>(n_total_jumps);
 }
 
 void ModelHawkesFixedExpKernLogLik::grad_i(const ulong sampled_i,
@@ -120,8 +121,8 @@ double ModelHawkesFixedExpKernLogLik::loss_and_grad(const ArrayDouble &coeffs,
                                    &ModelHawkesFixedExpKernLogLik::loss_and_grad_dim_i,
                                    this,
                                    coeffs, out);
-  out /= n_total_jumps;
-  return loss / n_total_jumps;
+  out /= static_cast<double>(n_total_jumps);
+  return loss / static_cast<double>(n_total_jumps);
 }
 
 double ModelHawkesFixedExpKernLogLik::hessian_norm(const ArrayDouble &coeffs,
@@ -134,7 +135,7 @@ double ModelHawkesFixedExpKernLogLik::hessian_norm(const ArrayDouble &coeffs,
                                    this,
                                    coeffs, vector);
 
-  return norm_sum / n_total_jumps;
+  return norm_sum / static_cast<double>(n_total_jumps);
 }
